+++ src/main/cpp/OpModes/AutomaticPath.cpp
@@ -4,7 +4,7 @@
 
 #include "OpModes/AutomaticPath.h"
 
-AutomaticPath::AutomaticPath(Robot * pRobot, std::vector<frc::Pose2d*> waypoints, Drivebase * MecanumDrive, Odometry * OdometryController, float speedMultiplier, float rotationMultiplier)
+AutomaticPath::AutomaticPath(Robot * pRobot, std::vector<frc::Pose2d*> waypoints, Drivebase * MecanumDrive, Odometry * OdometryController, double speedMultiplier, double rotationMultiplier)
 {
     name = "autoTrench";
     timesRun = 0;
@@ -36,16 +36,16 @@ bool AutomaticPath::Complete()
 }
 
 //Calculates distance to next waypoint
-float AutomaticPath::distanceToNextWaypoint(frc::Pose2d * waypoint)
+double AutomaticPath::distanceToNextWaypoint(frc::Pose2d * waypoint)
 {
-    float xCurrent = OdometryController->getX();
-    float yCurrent = OdometryController->getY();
-    float xWaypoint = (float)waypoint->Translation().X();
-    float yWaypoint = (float)waypoint->Translation().Y();
-    float dX = xCurrent - xWaypoint;
-    float dY = yCurrent - yWaypoint;
+    double xCurrent = OdometryController->getX();
+    double yCurrent = OdometryController->getY();
+    double xWaypoint = (double)waypoint->Translation().X();
+    double yWaypoint = (double)waypoint->Translation().Y();
+    double dX = xCurrent - xWaypoint;
+    double dY = yCurrent - yWaypoint;
     //the most efficient function in existance. Probably could do everything without square roots honestly
-    float distance = sqrt((dX * dX + dY * dY));
+    double distance = sqrt((dX * dX + dY * dY));
     PathDebug->PutNumber("Distance", distance);
     PathDebug->PutNumber("TargetX", xWaypoint);
     PathDebug->PutNumber("TargetY", yWaypoint);
@@ -55,18 +55,18 @@ float AutomaticPath::distanceToNextWaypoint(frc::Pose2d * waypoint)
 }
 
 //Return robot's angle to next waypoint using the nav gyro
-float AutomaticPath::angleToNextWaypoint(frc::Pose2d * waypoint)
+double AutomaticPath::angleToNextWaypoint(frc::Pose2d * waypoint)
 {
-    float xCurrent = OdometryController->getX();
-    float yCurrent = OdometryController->getY();
-    float xWaypoint = (float)waypoint->Translation().X();
-    float yWaypoint = (float)waypoint->Translation().Y();
-    float robotHeading = OdometryController->getYaw();
-    float alpha = atan2((yWaypoint - yCurrent), (xWaypoint - xCurrent)) * 180/3.14159265358979;
-    float beta = 90 - alpha;
-    float clockwiseAngleToWaypoint = robotHeading + beta;
-    float counterclockwise = 360 - clockwiseAngleToWaypoint;
-    float finalAngle = OdometryController->normalize360(counterclockwise);
+    double xCurrent = OdometryController->getX();
+    double yCurrent = OdometryController->getY();
+    double xWaypoint = (double)waypoint->Translation().X();
+    double yWaypoint = (double)waypoint->Translation().Y();
+    double robotHeading = OdometryController->getYaw();
+    double alpha = atan2((yWaypoint - yCurrent), (xWaypoint - xCurrent)) * 180/3.14159265358979;
+    double beta = 90 - alpha;
+    double clockwiseAngleToWaypoint = robotHeading + beta;
+    double counterclockwise = 360 - clockwiseAngleToWaypoint;
+    double finalAngle = OdometryController->normalize360(counterclockwise);
     PathDebug->PutNumber("Angle", finalAngle);
     return finalAngle;
     //I know this function is unreadable but it probably works so just ctrl-C ctrl-V
@@ -75,8 +75,8 @@ float AutomaticPath::angleToNextWaypoint(frc::Pose2d * waypoint)
 //Returns true if robot is pretty close to angle it needs to be at for the next waypoint
 bool AutomaticPath::atHeading(frc::Pose2d * waypoint)
 {
-    if((float)waypoint->Rotation().Degrees() > OdometryController->getYaw() - 2 &&
-       (float)waypoint->Rotation().Degrees() < OdometryController->getYaw() + 2)
+    if((double)waypoint->Rotation().Degrees() > OdometryController->getYaw() - 2 &&
+       (double)waypoint->Rotation().Degrees() < OdometryController->getYaw() + 2)
     {
         //Don't need to rotate (close enough to heading)
         return true;
@@ -92,12 +92,12 @@ bool AutomaticPath::atHeading(frc::Pose2d * waypoint)
 bool AutomaticPath::moveToNextWaypoint()
 {
     //When we reach the waypoint we delete it from the waypoint array.
-    float smallDistance1 = 0.05;
-    float smallDistance2 = 0.1; //this should be bigger than smallDistance1
-    float angle = angleToNextWaypoint(waypoints[waypointIndex]);
-    float speed = speedMultiplier;
+    double smallDistance1 = 0.05;
+    double smallDistance2 = 0.1; //this should be bigger than smallDistance1
+    double angle = angleToNextWaypoint(waypoints[waypointIndex]);
+    double speed = speedMultiplier;
     if(waypointIndex == (int)(waypoints.size() - 1)) {
-        float distance = distanceToNextWaypoint(waypoints[waypointIndex]);
+        double distance = distanceToNextWaypoint(waypoints[waypointIndex]);
         //When we start getting close to the last waypoint, slow down a bit
         if(distance < smallDistance2) {
             speed /= 2;
@@ -105,7 +105,7 @@ bool AutomaticPath::moveToNextWaypoint()
     }
 
     //If at the correct heading this becomes 0, if not it is scaled to the rotation speed
-    float rotate = (int)!atHeading(waypoints[waypointIndex]) * rotationMultiplier;
+    double rotate = (int)!atHeading(waypoints[waypointIndex]) * rotationMultiplier;
     MecanumDrive->PolarDrive(speed, angle, rotate, 0);
